refactor(sftp): move file src/controllers/files/sftp_utils.go -> src/utils/sftp_utils.go
